+++ CHANGELOG.md
@@ -8,7 +8,8 @@ All notable changes to this project will be documented in this file.
 - more examples
 - using a better product for product image classification test and example
 - add log4js
-- update translate service to 1810B (https://help.sap.com/viewer/2e6173bf645243bb9a88f0269250f3a2/1.0/en-US)
+- update translate service to [1810B](https://help.sap.com/viewer/2e6173bf645243bb9a88f0269250f3a2/1.0/en-US)
+- adjust detection_boxes for four people at human detection test
 
 ## 0.5.6